separate renderHTML function
